Mouse interactions corrected.
This commit is labeled as stable, but it still
has issues. (1) It needs to be formally tested and (2) It uses a hard-
coded initial display size--meaning it is possible for nodes to start
outside of the display window.

Some noteable changes include moving transform functions into the
graph dispatch so they are available to various callers and the
inclusion of a inverse view transform function. It would be
reasonable to automatically test that these functions are inverses.

This branch can be merged to development--but not yet to the main
branch.